==> src/1_pre_processing.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from utils import print_column_with_nan, clean_value, get_columns_with_nan

DATASET_PATH = '../data-new/acute_coronary_syndrome_db.xlsx'
df = pd.read_excel(DATASET_PATH, sheet_name=0, usecols=range(0, 74))

# Asumptions
NUMERICAL_FEATURES = ['EDAD','PESO Kg', 'ALTURA cm', 'TROPONINA MAX', 'CREAT', 'GLUCEMIA INGR', 'GB', 'TAS INGRESO', 'FC INGRESO']
NUMERICAL_CONTINOUS = NUMERICAL_FEATURES

CATEGORICAL_MULTI_CLASS_FEATURES = ['KILLIP Ingreso', 'FUNCION VENTRICULAR IZQ', 'DIAGNOSTICO', 'NUMERO VASOS', 'Peor KILLIP']
CATEGORICAL_BINARY_FEATURES = ['NUMERO VASOS', 'SEX0', 'HTA', 'DBT', 'DLP', 'TABAQ', 'ANTEC IAM / Angina inestable', 'ANTEC BY PASS', 'ANTEC ATC', 'ANTEC ACE', 'ANTEC INS RENAL', 'ACV TIA PREV', 'EPOC', 'CLAUD ITTE', 'BB PREV', ' IECA/AT2  PREV', 'B CA PREV', 'AAS PREV', 'HIPOGLUCEMIANTES', 'DIURETICOS', 'INSULINA', 'ECG INFRA ST', 'ECG SUPRA ST', 'ECG INV T', 'BRI', 'BRD', 'ECG MCP', 'ECG FA', 'tropst', 'TnT Ultrasensible', 'Elevacion troponina T', 'AAS', 'BB', 'IECA', 'AT2', 'Clopi - prasu - tica', 'HEP sc o iv', 'IIb IIIa', 'B CA', 'ESTATINAS', 'HIPOGLUC METFORM', 'HIPOGL SULFAN', 'HIPOGL GLITAZ', 'INSULINA', 'INS + HIPOGL', 'INOTROPICOS', 'PRUEBA FUNC.', 'PRUEBA FUNC DE ALTO RIESGO', 'CCG', 'TRONCO', 'ATC PRIMARIA', 'ATC INTRAHOSP', 'TROMBOL', 'CRM', 'IAM HOSP(SI INTERNO POR ANGINA) O REIAM', 'APIAM', 'ANGINA REFRACT o RECURR', ' ACV/TIA', 'SANGRADO MAYOR*', 'I RENAL AGUDA']

CATEGORY_FEATURES = CATEGORICAL_BINARY_FEATURES + CATEGORICAL_MULTI_CLASS_FEATURES

TARGET = "MUERTE HOSP"

print_column_with_nan(df)

Procedemos a ver cuales son las columnas tipo `object` y hacemos una limpieza de los datos erroneos, nulos o vacios. 

# Obtenemos las columnas de tipo object
object_columns = df.select_dtypes(include=['object']).columns

df[object_columns].info()

Vemos que tipo de datos unicos tenemos en cada columna

for column in object_columns:
  print(f"{column}: {df[column].unique()}")

Observamos tipos de lo mas variados: numeros, strings, dates, nan.

Vemos que hay valores string vacios, strings que representan una expresion numerica `<0,01`, guiones, texto aclaratorio, etc.
Tenemos que limpiar todos los simbolos y espacios en blanco. Tambien transformar los textos en valores que tengan sentido para esa feature. 

# Aplicamos la función solo a las columnas definidas en object_columns
df[object_columns] = df[object_columns].applymap(clean_value)

for column in object_columns:
  print(f"{column}: {df[column].unique()}")

Luego de limpiar los valores no deseados, pasamos todas las columnas a float (tipo de dato numérico que admite NaN) asi podemos empezar la imputación de los valores faltantes en el proximo notebook.

# Convertir las columnas a float para admitir NaN
df[object_columns] = df[object_columns].apply(pd.to_numeric, errors='coerce')

df.info()

print_column_with_nan(df, 2)

df.to_csv('../data-new/pre-processed-data.csv', index=False)
