==> src/4_feature_selection.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

DATASET_PATH = '../data-new/data-without-troponina.csv'
df = pd.read_csv(DATASET_PATH)

# Asumptions
NUMERICAL_FEATURES = ['EDAD','PESO Kg', 'ALTURA cm', 'CREAT', 'GLUCEMIA INGR', 'GB', 'TAS INGRESO', 'FC INGRESO']
NUMERICAL_CONTINOUS = NUMERICAL_FEATURES

CATEGORICAL_MULTI_CLASS_FEATURES = ['KILLIP Ingreso', 'FUNCION VENTRICULAR IZQ', 'DIAGNOSTICO', 'NUMERO VASOS', 'Peor KILLIP']
CATEGORICAL_BINARY_FEATURES = ['NUMERO VASOS', 'SEX0', 'HTA', 'DBT', 'DLP', 'TABAQ', 'ANTEC IAM / Angina inestable', 'ANTEC BY PASS', 'ANTEC ATC', 'ANTEC ACE', 'ANTEC INS RENAL', 'ACV TIA PREV', 'EPOC', 'CLAUD ITTE', 'BB PREV', ' IECA/AT2  PREV', 'B CA PREV', 'AAS PREV', 'HIPOGLUCEMIANTES', 'DIURETICOS', 'INSULINA', 'ECG INFRA ST', 'ECG SUPRA ST', 'ECG INV T', 'BRI', 'BRD', 'ECG MCP', 'ECG FA', 'tropst', 'TnT Ultrasensible', 'Elevacion troponina T', 'AAS', 'BB', 'IECA', 'AT2', 'Clopi - prasu - tica', 'HEP sc o iv', 'IIb IIIa', 'B CA', 'ESTATINAS', 'HIPOGLUC METFORM', 'HIPOGL SULFAN', 'HIPOGL GLITAZ', 'INSULINA', 'INS + HIPOGL', 'INOTROPICOS', 'PRUEBA FUNC.', 'PRUEBA FUNC DE ALTO RIESGO', 'CCG', 'TRONCO', 'ATC PRIMARIA', 'ATC INTRAHOSP', 'TROMBOL', 'CRM', 'IAM HOSP(SI INTERNO POR ANGINA) O REIAM', 'APIAM', 'ANGINA REFRACT o RECURR', ' ACV/TIA', 'SANGRADO MAYOR*', 'I RENAL AGUDA']

CATEGORY_FEATURES = CATEGORICAL_BINARY_FEATURES + CATEGORICAL_MULTI_CLASS_FEATURES

FEATURES = NUMERICAL_CONTINOUS + CATEGORY_FEATURES

TARGET = "MUERTE HOSP"

len(FEATURES)

Vamos a hacer una selección de features basado en la importancia de las features.

Para ello, vamos a implementar Información Mutua (MI). La información mutua mide la dependencia no lineal entre las variables independientes (features) y la variable objetivo. Se basa en la teoría de la información y cuantifica cuánto conocimiento sobre la variable objetivo se obtiene al conocer una característica específica.


Primero hacemos una imputacion de los valores faltantes provisoria para poder calcular la información mutua.

from sklearn.impute import SimpleImputer

df_copy = df.copy()

median_imputer = SimpleImputer(strategy='median')
mode_imputer = SimpleImputer(strategy='most_frequent')

median_imputer.fit(df_copy[NUMERICAL_FEATURES])
mode_imputer.fit(df_copy[CATEGORY_FEATURES])

df_copy[NUMERICAL_FEATURES] = median_imputer.transform(df_copy[NUMERICAL_FEATURES])
df_copy[CATEGORY_FEATURES] = mode_imputer.transform(df_copy[CATEGORY_FEATURES])


# Separar variables independientes (X) y variable objetivo (y)
X = df_copy[FEATURES]  # Variables predictoras
y = df_copy[TARGET]  # Variable objetivo

from sklearn.feature_selection import mutual_info_classif

# Calcular Información Mutua
mi_scores = mutual_info_classif(X, y, random_state=42)
mi_scores = pd.Series(mi_scores, index=X.columns)

# Mostrar la importancia de cada variable
print("Importancia de Información Mutua:")
print(mi_scores.sort_values(ascending=False))

# Seleccionar variables con Información Mutua mayor a 0.01
selected_mi_mask = mi_scores.apply(lambda x: x > 0.01)
selected_mi = mi_scores[selected_mi_mask]
print(selected_mi.sort_values(ascending=False))


from sklearn.ensemble import RandomForestClassifier

# Entrenar un modelo Random Forest para evaluar importancia de características
model_rf = RandomForestClassifier(n_estimators=100, random_state=42)
model_rf.fit(X, y)

# Obtener importancia de variables
rf_importances = pd.Series(model_rf.feature_importances_, index=X.columns)

# Seleccionar las variables más importantes
selected_rf_mask = rf_importances.apply(lambda x: x > 0.01)
selected_rf = rf_importances[selected_rf_mask]
print("Variables seleccionadas por Random Forest:\n", selected_rf.sort_values(ascending=False))


# Crear sets de características seleccionadas por cada método
features_mi = set(selected_mi.index)
features_rf = set(selected_rf.index)

# Comparar intersección (variables seleccionadas por varios métodos)
intersect_features = features_mi & features_rf
print("Variables seleccionadas por TODOS los métodos:\n", list(intersect_features))


Finalmente, vamos a seleccionar las variables que están en la intersección de los dos métodos.


selected_features = list(intersect_features)
print("Variables seleccionadas:\n", selected_features)

df_selected = df[selected_features + [TARGET]]

df_selected.info()

from utils import print_column_with_nan
print_column_with_nan(df_selected)

df_selected.to_csv('../data-new/data-selected-features.csv', index=False)
