==> src/7_variable_scaling.ipynb <==
import numpy as np
import pandas as pd

X_train = pd.read_csv('../data-new/train/X_train_imputed.csv')
y_train = pd.read_csv('../data-new/train/y_train.csv')
X_val = pd.read_csv('../data-new/val/X_val_imputed.csv')
y_val = pd.read_csv('../data-new/val/y_val.csv')
X_test = pd.read_csv('../data-new/test/X_test_imputed.csv')
y_test = pd.read_csv('../data-new/test/y_test.csv')

FEATURES = ['I RENAL AGUDA', 'NUMERO VASOS', 'SANGRADO MAYOR*', 'FUNCION VENTRICULAR IZQ', 'CRM', 'CREAT', 'KILLIP Ingreso', 'GB', 'Peor KILLIP', 'INOTROPICOS', 'TAS INGRESO', 'GLUCEMIA INGR']

NUMERICAL_FEATURES = ['CREAT', 'TAS INGRESO', 'GB', 'GLUCEMIA INGR']

CATEGORICAL_MULTI_CLASS_FEATURES = ['NUMERO VASOS', 'FUNCION VENTRICULAR IZQ', 'KILLIP Ingreso', 'Peor KILLIP', ]
CATEGORICAL_BINARY_FEATURES = ['I RENAL AGUDA', 'SANGRADO MAYOR*', 'CRM',  'INOTROPICOS',]

CATEGORY_FEATURES = CATEGORICAL_BINARY_FEATURES + CATEGORICAL_MULTI_CLASS_FEATURES

TARGET = "MUERTE HOSP"

len(FEATURES)

Realizamos un escalado de las variables numéricas

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()

X_train[NUMERICAL_FEATURES] = scaler.fit_transform(X_train[NUMERICAL_FEATURES])
X_val[NUMERICAL_FEATURES] = scaler.transform(X_val[NUMERICAL_FEATURES])
X_test[NUMERICAL_FEATURES] = scaler.transform(X_test[NUMERICAL_FEATURES])

X_train.to_csv('../data-new/train/X_train_scaled.csv', index=False)
X_val.to_csv('../data-new/val/X_val_scaled.csv', index=False)
X_test.to_csv('../data-new/test/X_test_scaled.csv', index=False)
